Changed QuickShot system slightly: now it takes a screenshot _before_ area selection. This ensures the active window does not change. However, this has made clear some small artifacts as a result of pixbuf capture - I will need to look into capturing via a different system e.g. Compiz (if available) or X11.
Also modified area selection: a rather nice little function was added to make the border colour use a system UI colour, provided that the contrast with the BG is high enough.

Fixed one or two bugs introduced in the last commit too: PyNotify initiation was missed out (causing the rest of the pynotify code to fail), and there was a small inconsistency included in the FileConfigManager defaults system.